requirements.txt
tensorflow>=2.13.0
numpy>=1.23.5
pandas>=2.0.3
matplotlib>=3.7.2
scikit-learn>=1.3.0
pillow>=10.0.0
seaborn>=0.12.2
tqdm>=4.65.0 